Modified SP scripts, Create schema script and schema diagram

+++ db/sql-scripts/Create_SP_Employee.sql
@@ -1,7 +1,11 @@
 -- STORED PROCEDURES FOR EmployeeService
 
 
--- Create new support doc
+-- -----------------------------------------------------
+-- procedure create_support_doc
+--    - create a support doc for the given employee,
+--    - only if the employee and type exist
+-- -----------------------------------------------------
 DROP PROCEDURE IF EXISTS create_support_doc;
 
 DELIMITER //
@@ -15,16 +19,32 @@ CREATE PROCEDURE `create_support_doc` (IN employee_id INT
 , IN description VARCHAR(512)
 )
 BEGIN
-    INSERT INTO `SUPPORT_DOC` (EMPLOYEE_ID, TYPE_ID, CREATED_DATE, DUE_DATE, EXPIRY_DATE, PATH, DESCRIPTION)
-    VALUES (employee_id, type_id, created_date, due_date, expiry_date, path, description);
+    DECLARE typeChecker INT;
+    DECLARE emplChecker INT;
+    SET typeChecker = 0;
+    SET emplChecker = 0;
+    SELECT COUNT(TYPE_ID) INTO typeChecker FROM `TYPE` WHERE `TYPE`.TYPE_ID = type_id;
+    SELECT COUNT(EMPLOYEE_ID) INTO emplChecker FROM `HR_RECORD` WHERE `HR_RECORD`.EMPLOYEE_ID = employee_id;
+
+    IF typeCheckcer = 0 THEN
+      SIGNAL SQLSTATE '45000' SET MESSAGE_TEXT = 'Document type does not exist.';
+    ELSEIF emplChecker = 0 THEN
+      SIGNAL SQLSTATE '45000' SET MESSAGE_TEXT = 'Employee does not exist.'
+    ELSE
+      INSERT INTO `SUPPORT_DOC` (EMPLOYEE_ID, TYPE_ID, CREATED_DATE, DUE_DATE, EXPIRY_DATE, PATH, DESCRIPTION)
+      VALUES (employee_id, type_id, created_date, due_date, expiry_date, path, description);
+    END IF;
 END //
 
 DELIMITER ;
 
 
--- Create an employee in the database
--- It will be automatically assigned an employee_id
--- Version = 0 indicating its a new employee
+-- -----------------------------------------------------
+-- procedure create_employee
+--    - create an employee, provided SIN and email are
+--    - not already in use
+--    - Version = 0 to indicate a new employee
+-- -----------------------------------------------------
 DROP PROCEDURE IF EXISTS create_employee;
 
 DELIMITER //
@@ -49,18 +69,42 @@ CREATE PROCEDURE `create_employee`(IN created_by_id INT
 , IN phone_number VARCHAR(45)
 )
 BEGIN
-    INSERT INTO HR_RECORD (
+    DECLARE sinChecker INT;
+    DECLARE emailChecker INT;
+
+    SET sinChecker = 0;
+    SET emailChecker = 0;
+
+    SELECT COUNT(SIN) INTO sinChecker
+    FROM `HR_RECORD`
+    WHERE `HR_RECORD`.SIN = SIN;
+
+    SELECT COUNT(EMAIL) INTO emailChecker
+    FROM `HR_RECORD`
+    WHERE `HR_RECORD`.EMAIL = email;
+
+    IF sinChecker > 0 THEN
+      SIGNAL SQLSTATE '45000' SET MESSAGE_TEXT = 'SIN already in use.';
+    ELSEIF emailChecker > 0 THEN
+      SIGNAL SQLSTATE '45000' SET MESSAGE_TEXT = 'Email already in use.';
+    ELSE
+      INSERT INTO HR_RECORD (
       VERSION, CREATED_BY, ROLE, SIN, EMAIL, FIRST_NAME, LAST_NAME, ADDRESS, BIRTHDATE, VACATION_DAYS,
       REMAINING_VACATION_DAYS, FTE, STATUS, PASSWORD, SALARY, DATE_JOINED, ADMIN_LEVEL, CREATED_DATE, PHONE_NUMBER)
-    VALUES (
+      VALUES (
       0, created_by_id, role, SIN, email, first_name, last_name, address, birthdate, vacation_days,
       remaining_vacation_days, FTE, status, password, salary, date_joined, admin_level, created_date, phone_number);
+    END IF;
 END //
 
 DELIMITER ;
 
 
--- Create performance for given employee
+-- -----------------------------------------------------
+-- procedure create_employee_performance
+--    - create a performance record for an employee,
+--    - provided that the employee exists
+-- -----------------------------------------------------
 DROP PROCEDURE IF EXISTS create_employee_performance;
 
 DELIMITER //
@@ -70,14 +114,31 @@ CREATE PROCEDURE `create_employee_performance` (IN employee_id INT
 , IN status TINYINT
 )
 BEGIN
-    INSERT INTO `PERFORMANCE` (`DATE`, STATUS, EMPLOYEE_ID)
-    VALUES (created_date, status, employee_id);
+    DECLARE checker INT;
+
+    SET checker = 0;
+
+    SELECT COUNT(EMPLOYEE_ID) INTO checker
+    FROM `HR_RECORD`
+    WHERE `HR_RECORD`.EMPLOYEE_ID = employee_id;
+
+    IF checker = 0 THEN
+      SIGNAL SQLSTATE '45000' SET MESSAGE_TEXT = 'Employee does not exist.';
+    ELSE
+      INSERT INTO `PERFORMANCE` (`DATE`, STATUS, EMPLOYEE_ID)
+      VALUES (created_date, status, employee_id);
+    END IF;
 END //
 
 DELIMITER ;
 
 
--- Create vacation for given employee
+-- -----------------------------------------------------
+-- procedure create_employee_vacation
+--    - create a vacation request for an employee,
+--    - assigned to another employee to aprove, provided
+--    - they both exist
+-- -----------------------------------------------------
 DROP PROCEDURE IF EXISTS create_employee_vacation;
 
 DELIMITER //
@@ -89,18 +150,39 @@ CREATE PROCEDURE `create_employee_vacation` (IN employee_id INT
 , IN created_date BIGINT
 )
 BEGIN
-    INSERT INTO `VACATION_REQUEST` (EMPLOYEE_ID, APPROVER_ID, REQUESTED_DAYS, REQUEST_STATUS, `DATE`)
-    VALUES (employee_id, approver_id, requested_days, request_status, created_date);
+    DECLARE emplChecker INT;
+    DECLARE approvChecker INT;
+
+    SET emplChecker = 0;
+    SET approvChecker = 0;
+
+    SELECT COUNT(EMPLOYEE_ID) INTO emplChecker
+    FROM `HR_RECORD`
+    WHERE `HR_RECORD`.EMPLOYEE_ID = employee_id;
+
+    SELECT COUNT(EMPLOYEE_ID) INTO approvChecker
+    FROM `HR_RECORD`
+    WHERE `HR_RECORD`.EMPLOYEE_ID = approver_id;
+
+    IF emplChecker = 0 THEN
+      SIGNAL SQLSTATE '45000' SET MESSAGE_TEXT = 'Employee does not exist.';
+    ELSEIF approvChecker = 0 THEN
+      SIGNAL SQLSTATE '45000' SET MESSAGE_TEXT = 'Approver employee does not exist.';
+    ELSE
+      INSERT INTO `VACATION_REQUEST` (EMPLOYEE_ID, APPROVER_ID, REQUESTED_DAYS, REQUEST_STATUS, `DATE`)
+      VALUES (employee_id, approver_id, requested_days, request_status, created_date);
+    END IF;
 END //
 
 DELIMITER ;
 
 
--- Update an employee in the database
--- Essentially a new row with incremented version and updated values
---
--- Deleting an employee can also use this by setting the status
--- to "0 = Inactive"
+-- -----------------------------------------------------
+-- procedure update_employee
+--    - update an employee, provided that employee exists
+--    - also validate that the new SIN and EMAIL are not
+--    - used by other employees
+-- -----------------------------------------------------
 DROP PROCEDURE IF EXISTS update_employee;
 
 DELIMITER //
@@ -127,69 +209,148 @@ CREATE PROCEDURE `update_employee`(IN employee_id INT
 , IN phone_number VARCHAR(45)
 )
 BEGIN
-    INSERT INTO HR_RECORD (
+    DECLARE sinChecker INT;
+    DECLARE emailChecker INT;
+    DECLARE emplChecker INT;
+
+    SET sinChecker = 0;
+    SET emailChecker = 0;
+    SET emplChecker = 0;
+
+    SELECT COUNT(SIN) INTO sinChecker
+    FROM `HR_RECORD`
+    WHERE `HR_RECORD`.SIN = SIN AND `HR_RECORD`.employee_id <> employee_id;
+
+    SELECT COUNT(EMAIL) INTO emailChecker
+    FROM `HR_RECORD`
+    WHERE `HR_RECORD`.EMAIL = email AND `HR_RECORD`.employee_id <> employee_id;
+
+    SELECT COUNT(EMPLOYEE_ID) INTO emplChecker
+    FROM `HR_RECORD`
+    WHERE `HR_RECORD`.EMPLOYEE_ID = employee_id;
+
+    IF sinChecker > 0 THEN
+      SIGNAL SQLSTATE '45000' SET MESSAGE_TEXT = 'SIN already in use.';
+    ELSEIF emailChecker > 0 THEN
+      SIGNAL SQLSTATE '45000' SET MESSAGE_TEXT = 'Email already in use.';
+    ELSEIF employeeChecker = 0 THEN
+      SIGNAL SQLSTATE '45000' SET MESSAGE_TEXT = 'Employee does not exist';
+    ELSE
+      INSERT INTO HR_RECORD (
       EMPLOYEE_ID, VERSION, CREATED_BY, ROLE, SIN, EMAIL, FIRST_NAME, LAST_NAME, ADDRESS, BIRTHDATE, VACATION_DAYS,
       REMAINING_VACATION_DAYS, FTE, STATUS, PASSWORD, SALARY, DATE_JOINED, ADMIN_LEVEL, CREATED_DATE, PHONE_NUMBER)
-    VALUES (
+      VALUES (
       employee_id, version, created_by_id, role, SIN, email, first_name, last_name, address, birthdate, vacation_days,
       remaining_vacation_days, FTE, status, password, salary, date_joined, admin_level, created_date, phone_number);
+    END IF;
 END //
 
 DELIMITER ;
 
 
--- Get documents of an employee, given employee id
+-- -----------------------------------------------------
+-- procedure get_employee_docs
+--    - get the documents of an employee, provided
+--    - the employee exists
+-- -----------------------------------------------------
 DROP PROCEDURE IF EXISTS get_employee_docs;
 
 DELIMITER //
 
-CREATE PROCEDURE `get_employee_docs`(IN id INT)
+CREATE PROCEDURE `get_employee_docs`(IN employee_id INT)
 BEGIN
-  SELECT *
-  FROM SUPPORT_DOC
-  WHERE EMPLOYEE_ID = id;
+  DECLARE checker INT;
+
+  SET checker = 0;
+
+  SELECT COUNT(EMPLOYEE_ID) INTO checker
+  FROM `HR_RECORD`
+  WHERE `HR_RECORD`.EMPLOYEE_ID = employee_id;
+
+  IF checker = 0 THEN
+    SIGNAL SQLSTATE '45000' SET MESSAGE_TEXT = 'Employee does not exist.';
+  ELSE
+    SELECT *
+    FROM SUPPORT_DOC
+    WHERE EMPLOYEE_ID = employee_id;
+  END IF;
 END //
 
 DELIMITER ;
 
 
--- Get latest version of an employee, given an employee ID
+-- -----------------------------------------------------
+-- procedure get_employee
+--    - get the latest version of the employee,
+--    - provided the employee exists
+-- -----------------------------------------------------
 DROP PROCEDURE IF EXISTS get_employee;
 
 DELIMITER //
 
-CREATE PROCEDURE `get_employee`(IN id INT)
+CREATE PROCEDURE `get_employee`(IN employee_id INT)
 BEGIN
-  SELECT *
-  FROM HR_RECORD hr1
-  INNER JOIN (
-    SELECT EMPLOYEE_ID, MAX(VERSION)
-    FROM HR_RECORD
-    GROUP BY EMPLOYEE_ID
-  ) hr2 ON hr1.EMPLOYEE_ID = hr2.EMPLOYEE_ID AND hr1.VERSION = hr2.VERSION
-  WHERE hr1.EMPLOYEE_ID = id;
+  DECLARE checker INT;
+
+  SET checker = 0;
+
+  SELECT COUNT(EMPLOYEE_ID) INTO checker
+  FROM `HR_RECORD`
+  WHERE `HR_RECORD`.EMPLOYEE_ID = employee_id;
+
+  IF checker = 0 THEN
+    SIGNAL SQLSTATE '45000' SET MESSAGE_TEXT = 'Employee does not exist.';
+  ELSE
+    SELECT *
+    FROM HR_RECORD hr1
+    INNER JOIN (
+      SELECT EMPLOYEE_ID, MAX(VERSION)
+      FROM HR_RECORD
+      GROUP BY EMPLOYEE_ID
+    ) hr2 ON hr1.EMPLOYEE_ID = hr2.EMPLOYEE_ID AND hr1.VERSION = hr2.VERSION
+    WHERE hr1.EMPLOYEE_ID = id;
+  END IF;
 END //
 
 DELIMITER ;
 
 
--- Get history of an employee, given an employee ID
+-- -----------------------------------------------------
+-- procedure get_employee_history
+--    - get the history of the employee,
+--    - provided the employee exists
+-- -----------------------------------------------------
 DROP PROCEDURE IF EXISTS get_employee_history;
 
 DELIMITER //
 
-CREATE PROCEDURE `get_employee_history`(IN id INT)
+CREATE PROCEDURE `get_employee_history`(IN employee_id INT)
 BEGIN
-  SELECT *
-  FROM HR_RECORD
-  WHERE hr1.EMPLOYEE_ID = id
-  ORDER BY VERSION DESC;
+  DECLARE checker INT;
+
+  SET checker = 0;
+
+  SELECT COUNT(EMPLOYEE_ID) INTO checker
+  FROM `HR_RECORD`
+  WHERE `HR_RECORD`.EMPLOYEE_ID = employee_id;
+
+  IF checker = 0 THEN
+    SIGNAL SQLSTATE '45000' SET MESSAGE_TEXT = 'Employee does not exist.';
+  ELSE
+    SELECT *
+    FROM HR_RECORD
+    WHERE hr1.EMPLOYEE_ID = employee_id
+    ORDER BY VERSION DESC;
+  END IF;
 END //
 
 DELIMITER ;
 
 
--- Get latest versions of all employees
+-- -----------------------------------------------------
+-- procedure get_all_employees
+--    - get the latest version of all employees
+-- -----------------------------------------------------
 DROP PROCEDURE IF EXISTS get_all_employees;
 
 DELIMITER //
@@ -208,14 +369,28 @@ END //
 DELIMITER ;
 
 
--- Get latest versions of all employees under the given
--- manager ID
+-- -----------------------------------------------------
+-- procedure get_manager_employees
+--    - get the employees reporting under the given
+--    - manager, provided the manager exists
+-- -----------------------------------------------------
 DROP PROCEDURE IF EXISTS get_manager_employees;
 
 DELIMITER //
 
-CREATE PROCEDURE `get_manager_employees` (IN id INT)
+CREATE PROCEDURE `get_manager_employees` (IN manager_id INT)
 BEGIN
+  DECLARE checker INT;
+
+  SET checker = 0;
+
+  SELECT COUNT(EMPLOYEE_ID) INTO checker
+  FROM `HR_RECORD`
+  WHERE `HR_RECORD`.EMPLOYEE_ID = manager_id;
+
+  IF checker = 0 THEN
+    SIGNAL SQLSTATE '45000' SET MESSAGE_TEXT = 'Manager does not exist.';
+  ELSE
     SELECT *
     FROM HR_RECORD hr1
     INNER JOIN (
@@ -226,20 +401,36 @@ BEGIN
     WHERE hr1.EMPLOYEE_ID in (
       SELECT m.EMPLOYEE_ID
       FROM MANAGER_EMPLOYEE m
-      WHERE m.MANAGER_ID = id
+      WHERE m.MANAGER_ID = manager_id
     );
+  END IF;
 END //
 
 DELIMITER ;
 
 
--- Get performance reviews and its respective sections for a given employee ID
+-- -----------------------------------------------------
+-- procedure get_performance_reviews
+--    - get the performance reviews for a given employee,
+--    - provided the employee exists
+-- -----------------------------------------------------
 DROP PROCEDURE IF EXISTS get_performance_reviews;
 
 DELIMITER //
 
-CREATE PROCEDURE `get_performance_reviews` (IN id INT)
+CREATE PROCEDURE `get_performance_reviews` (IN employee_id INT)
 BEGIN
+  DECLARE checker INT;
+
+  SET checker = 0;
+
+  SELECT COUNT(EMPLOYEE_ID) INTO checker
+  FROM `HR_RECORD`
+  WHERE `HR_RECORD`.EMPLOYEE_ID = employee_id;
+
+  IF checker = 0 THEN
+    SIGNAL SQLSTATE '45000' SET MESSAGE_TEXT = 'Employee does not exist.';
+  ELSE
     SELECT *
     FROM PERFORMANCE p
     LEFT JOIN JABC_GOAL jg ON p.PERFORMANCE_ID = jg.PERFORMANCE_ID
@@ -247,28 +438,49 @@ BEGIN
     LEFT JOIN OBJECTIVE o ON p.PERFORMANCE_ID = o.PERFORMANCE_ID
     LEFT JOIN DEVELOPMENT_GOAL dg ON p.PERFORMANCE_ID = dg.PERFORMANCE_ID
     LEFT JOIN COMMENT c ON p.PERFORMANCE_ID = c.PERFORMANCE_ID
-    WHERE p.EMPLOYEE_ID = id;
+    WHERE p.EMPLOYEE_ID = employee_id;
+  END IF;
 END //
 
 DELIMITER ;
 
 
--- Get all vacation requests for a given employee ID
+-- -----------------------------------------------------
+-- procedure get_employee_vacation
+--    - get the vacation requests for a given employee,
+--    - provided the employee exists
+-- -----------------------------------------------------
 DROP PROCEDURE IF EXISTS get_employee_vacation;
 
 DELIMITER //
 
 CREATE PROCEDURE `get_employee_vacation` (IN id INT)
 BEGIN
+  DECLARE checker INT;
+
+  SET checker = 0;
+
+  SELECT COUNT(EMPLOYEE_ID) INTO checker
+  FROM `HR_RECORD`
+  WHERE `HR_RECORD`.EMPLOYEE_ID = employee_id;
+
+  IF checker = 0 THEN
+    SIGNAL SQLSTATE '45000' SET MESSAGE_TEXT = 'Employee does not exist.';
+  ELSE
     SELECT *
     FROM VACATION_REQUEST
     WHERE EMPLOYEE_ID = id;
+  END IF;
 END //
 
 DELIMITER ;
 
 
--- Link employee with manager
+-- -----------------------------------------------------
+-- procedure link_employee_manager
+--    - link the employee to a manager,
+--    - provided both employee and manager exist
+-- -----------------------------------------------------
 DROP PROCEDURE IF EXISTS link_employee_manager;
 
 DELIMITER //
@@ -276,14 +488,38 @@ DELIMITER //
 CREATE PROCEDURE `link_employee_manager` (IN e_id INT
 , IN m_id INT)
 BEGIN
+  DECLARE emplChecker INT;
+  DECLARE managChecker INT;
+
+  SET emplChecker = 0;
+  SET managChecker = 0;
+
+  SELECT COUNT(EMPLOYEE_ID) INTO emplChecker
+  FROM `HR_RECORD`
+  WHERE `HR_RECORD`.EMPLOYEE_ID = e_id;
+
+  SELECT COUNT(EMPLOYEE_ID) INTO managChecker
+  FROM `HR_RECORD`
+  WHERE `HR_RECORD`.EMPLOYEE_ID = m_id;
+
+  IF emplChecker = 0 THEN
+    SIGNAL SQLSTATE '45000' SET MESSAGE_TEXT = 'Employee does not exist.';
+  ELSEIF managChecker = 0 THEN
+    SIGNAL SQLSTATE '45000' SET MESSAGE_TEXT = 'Manager employee does not exist.';
+  ELSE
     INSERT INTO MANAGER_EMPLOYEE (MANAGER_ID, EMPLOYEE_ID)
     VALUES (e_id, m_id);
+  END IF;
 END //
 
 DELIMITER ;
 
 
--- Link employee with manager
+-- -----------------------------------------------------
+-- procedure unlink_employee_manager
+--    - unlink the employee to a manager,
+--    - provided both employee and manager exist
+-- -----------------------------------------------------
 DROP PROCEDURE IF EXISTS unlink_employee_manager;
 
 DELIMITER //
@@ -291,8 +527,28 @@ DELIMITER //
 CREATE PROCEDURE `unlink_employee_manager` (IN e_id INT
 , IN m_id INT)
 BEGIN
+  DECLARE emplChecker INT;
+  DECLARE managChecker INT;
+
+  SET emplChecker = 0;
+  SET managChecker = 0;
+
+  SELECT COUNT(EMPLOYEE_ID) INTO emplChecker
+  FROM `HR_RECORD`
+  WHERE `HR_RECORD`.EMPLOYEE_ID = e_id;
+
+  SELECT COUNT(EMPLOYEE_ID) INTO managChecker
+  FROM `HR_RECORD`
+  WHERE `HR_RECORD`.EMPLOYEE_ID = m_id;
+
+  IF emplChecker = 0 THEN
+    SIGNAL SQLSTATE '45000' SET MESSAGE_TEXT = 'Employee does not exist.';
+  ELSEIF managChecker = 0 THEN
+    SIGNAL SQLSTATE '45000' SET MESSAGE_TEXT = 'Manager employee does not exist.';
+  ELSE
     DELETE FROM MANAGER_EMPLOYEE
     WHERE MANAGER_ID = m_id AND EMPLOYEE_ID = e_id;
+  END IF;
 END //
 
 DELIMITER ;
